+++ java/tech/datatype/LongReaderIter.java
@@ -5,8 +5,8 @@ import clojure.lang.Keyword;
 
 public class LongReaderIter implements IOBase, LongIter
 {
-  int idx;
-  int num_elems;
+  long idx;
+  long num_elems;
   LongReader reader;
   public LongReaderIter(LongReader _reader)
   {
@@ -15,7 +15,7 @@ public class LongReaderIter implements IOBase, LongIter
     reader = _reader;
   }
   public Keyword getDatatype() { return reader.getDatatype(); }
-  public int size() { return num_elems - idx; }
+  public long size() { return num_elems - idx; }
   public boolean hasNext() { return idx < num_elems; }
   public long nextLong() {
     long retval = reader.read(idx);
