+++ java/tech/datatype/MutableRemove.java
@@ -4,5 +4,5 @@ import java.nio.IntBuffer;
 
 public interface MutableRemove extends IOBase
 {
-  void remove(int idx);
+  void remove(long idx);
 }
